вынесла модули лайков, скрытия формы, апи, корректировки тегов
